Properly order input events to snapshot
When `snapshot` receives updates both to its input cell and stream, the old
cell value is used to generate the output event.

This is achieved by registering "finalizers" in a thread-local storage, to be
called at the end of committing a transaction. When a cell update is received,
snapshot does not update immediately but stores the new value in a temporary
container, while retaining the current value first. At the end of the
transaction a finalizer updates the current value to the new one.