Lifting of up to quaternary functions
This is a preliminary implementation to provide the practical functionality of
n-ary lifting for the most common cases. The type system did not allow to
implement this as a regular function/method without losing type inference.
Therefore a new macro `lift!` was implemented to dispatch to the
implementations for different arity.

Breaking change:

- `lift2` is no longer exported at the crate level, but in a module `lift`. It
  is recommended to use the macro `lift!` instead.