Allow transactions to be nested
Parent transactions are now stored in the stack frame of a new transaction.
They are restored as the current transaction, after the child transaction is
finished. This allows arbitrary nesting of transactions.

Note that there is still no true concurrency as only one thread at a time can
be executing transactions.