Wrap Signal<ReadOnly> as SignalMut
With the new interface there is less risk of using the interface in a manner
that is inconsistent with the FRP abstraction. The SignalMut is much more
restricted to guarantee this.

Note that one can still use Signal<ReadOnly> or Signal<Mutex> to side-step
this. But this is no longer part of the API and highly discouraged.