Cache return value of a signal function
Within a transaction, a signal should yield the same value, no matter how often
it is sampled for various reasons. For functions with side-effects (e.g. random
number generation, system time) one cannot guarantee that they return the same
value. Therefore the result of the function is cached for the duration of a
transaction. The cache is reset at the end of the transaction, so that the
signal function may return a different value later.