Revise the transaction system
For coalesce to work properly, there need to be two stages of
end-of-transaction callback execution. In the first stage, callbacks that might
spawn new events are called. This stage is repeated until no further callbacks
remain. (Note that each callback execution can register new callbacks.) In the
second stage actual finalizers are called. These do nothing but update pending
values. Their side-effects should always be local, so that this actually
finalizes the transaction.