Fix private_in_public warnings
This fixes the "warning: private type in public interface" warnings.
Rustc tells us that "this was previously accepted by the compiler but is
being phased out; it will become a hard error in a future release!".

Please note that this means that the `FuncSignal` struct (but not its
members will be public).